+++ packages/extended-yo-generator-test/src/index.ts
@@ -1,0 +1,2 @@
+export { IRunContext } from "./IRunContext";
+export { TestContext } from "./TestContext";