+++ packages/extended-yo-generator-test/src/tests/main.test.ts
@@ -1,0 +1,6 @@
+suite(
+    "ExtendedYoGeneratorTest",
+    () =>
+    {
+        require("./TestContext.test");
+    });