+++ packages/extended-yo-generator/.eslintrc.js
@@ -1,0 +1,15 @@
+const Path = require("path");
+
+module.exports = {
+    parser: "@typescript-eslint/parser",
+    extends: [
+        Path.join("..", "..", ".eslintrc.js")
+    ],
+    parserOptions: {
+        project: [
+            Path.join(__dirname, "tsconfig.json"),
+            Path.join(__dirname, "tsconfig.eslint.json"),
+            Path.join(__dirname, "src", "tests", "tsconfig.json")
+        ]
+    }
+};