+++ packages/extended-yo-generator/src/tests/Generator/index.ts
@@ -1,0 +1,15 @@
+import { TestContext } from "@manuth/extended-yo-generator-test";
+import { ExtendedGeneratorTests } from "./GeneratorTests.test";
+import { ITestOptions } from "./ITestOptions";
+import { TestGenerator } from "./TestGenerator/TestGenerator";
+
+/**
+ * Registers the generator-tests.
+ *
+ * @param context
+ * The context to use.
+ */
+export function GeneratorTests(context: TestContext<TestGenerator, ITestOptions>): void
+{
+    ExtendedGeneratorTests(context);
+}